updated img for readme, added numpy_utils file
