+++ src/cpp/src/text2image/schedulers/lcm.cpp
@@ -8,6 +8,7 @@
 
 #include "text2image/schedulers/lcm.hpp"
 #include "utils.hpp"
+#include "text2image/numpy_utils.hpp"
 
 
 namespace ov {
@@ -63,7 +64,7 @@ LCMScheduler::LCMScheduler(const Config& scheduler_config)
         float start = std::sqrt(m_config.beta_start);
         float end = std::sqrt(m_config.beta_end);
 
-        using utils::linspace;
+        using numpy_utils::linspace;
         std::vector<float> temp = linspace<float, float>(start, end, m_config.num_train_timesteps, true);
         for (float b : temp) {
             betas.push_back(b * b);
@@ -105,7 +106,7 @@ void LCMScheduler::set_timesteps(size_t num_inference_steps) {
     // LCM Inference Steps Schedule
     std::reverse(lcm_origin_timesteps.begin(),lcm_origin_timesteps.end());
 
-    using utils::linspace;
+    using numpy_utils::linspace;
     // v1. based on https://github.com/huggingface/diffusers/blame/2a7f43a73bda387385a47a15d7b6fe9be9c65eb2/src/diffusers/schedulers/scheduling_lcm.py#L387
     std::vector<size_t> inference_indices = linspace<size_t, float>(0, origin_timesteps_size, m_num_inference_steps);
     for (size_t i : inference_indices){
