+++ src/cpp/src/utils.hpp
@@ -70,28 +70,6 @@ ov::genai::StreamerVariant get_streamer_from_map(const ov::AnyMap& config_map);
 
 ov::genai::OptionalGenerationConfig get_config_from_map(const ov::AnyMap& config_map);
 
-// https://gist.github.com/lorenzoriano/5414671
-template <typename T, typename U>
-std::vector<T> linspace(U start, U end, size_t num, bool endpoint = false) {
-    std::vector<T> indices;
-    if (num != 0) {
-        if (num == 1)
-            indices.push_back(static_cast<T>(start));
-        else {
-            if (endpoint)
-                --num;
-
-            U delta = (end - start) / static_cast<U>(num);
-            for (size_t i = 0; i < num; i++)
-                indices.push_back(static_cast<T>(start + delta * i));
-
-            if (endpoint)
-                indices.push_back(static_cast<T>(end));
-        }
-    }
-    return indices;
-}
-
 }  // namespace utils
 }  // namespace genai
 }  // namespace ov
